Refactor Windows DnsResolver logic into DnsResolverPal (PAL pattern)
Move all platform-specific logic from DnsResolver.Windows.cs and
DnsResolver.WindowsAsync.cs into a new DnsResolverPal.Windows.cs static
PAL class, mirroring the existing NameResolutionPal pattern. The
cross-platform Resolve*Core methods remain on DnsResolver and delegate to
the PAL, providing a seam for future instrumentation and telemetry.
